scrapper/requirements.txt
requests==2.24.0
beautifulsoup4==4.9.1
python-dotenv==0.14.0